Refactor filename by removing 'Application'
